@@ -14,20 +14,20 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  bloc: ^6.0.0
-  remote_state: ^1.4.0
-  get_it:
-  get_it_mixin:
-  freezed_annotation:
-  dio: ^3.0.9
-  animations: ^1.1.2
+  bloc: ^7.0.0
+  remote_state: ^2.0.0
+  get_it: ^7.1.3
+  get_it_mixin: ^3.1.1
+  freezed_annotation: ^0.14.2
+  dio: ^4.0.0
+  animations: ^2.0.0
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
